fix leak in the test
