+++ src/index.ts
@@ -1,26 +1,5 @@
 import cluster from "cluster";
-import express from "express";
-import { evaluateScript } from './evaluate-script'
-import { base64Decode, bodyToEvent } from "./utils";
-
-export function buildApp() {
-  const app = express();
-  app.use(express.json({limit: '32mb'}))
-  app.post("/graphql-worker", async (req, res, next) => {
-    try {
-      const source = base64Decode(req.body.source) || req.body.source
-      const runner = evaluateScript(source)
-      const result = await runner(bodyToEvent(req.body));
-      if(result === undefined && req.body.resolver !== '$webhook') {
-        res.status(400)
-      }
-      res.json(result)
-    } catch(err) {
-      next(err)
-    }
-  })
-  return app;
-}
+import { buildApp } from "./buildApp"
 
 async function startServer() {
   const app = buildApp()
