--- src/utils.ts
@@ -1,25 +1,0 @@
-import atob from "atob";
-import btoa from "btoa";
-import { GraphQLEventFields } from '@slash-graphql/lambda-types'
-
-export function bodyToEvent(b: any): GraphQLEventFields {
-  return {
-    type: b.resolver,
-    parents: b.parents || null,
-    args: b.args || {},
-    authHeader: b.authHeader,
-    event: b.event || {},
-    info: b.info || null,
-  }
-}
-
-export function base64Decode(str: string) {
-  try {
-    const original = str.trim();
-    const decoded = atob(original);
-    return btoa(decoded) === original ? decoded : "";
-  } catch (err) {
-    console.error(err);
-    return "";
-  }
-}
